% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.r
\name{go}
\alias{go}
\title{Start the SHINY server}
\usage{
go(appDir = file.path(path.package(package = "jamestest"), "chart"), ...)
}
\arguments{
\item{appDir}{See \code{runApp}}

\item{\dots}{Passed down to \code{runApp}}
}
\description{
Start the SHINY server
}
