% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\docType{data}
\name{installed.cabinets}
\alias{installed.cabinets}
\title{Installed user cabinets}
\format{A list of \code{cabinet} objects. The \code{cabinet} object is
defined in the \code{minihealth} package.}
\usage{
installed.cabinets
}
\description{
A data set containing several user cabinets in which edits can be
made.
}
\keyword{datasets}
