% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_chart_controls.R
\name{update_chart_controls}
\alias{update_chart_controls}
\title{Update the user interface when a new chart is selected}
\usage{
update_chart_controls(session, chartgrp, chartcode, ga = NULL)
}
\arguments{
\item{session}{Shiny session object}

\item{chartgrp}{The chart group: \code{'nl2010'}, \code{'preterm'}, \code{'who'}
or \code{character(0)}}

\item{chartcode}{The chart code}

\item{ga}{Optionally, gestional age (weeks) for setting week slider
for non-preterms}
}
\value{
This function is called for its side effects
}
\description{
This function updates chart controls given the chart code.
The function updates the age group setting, the side radio button,
the etnicity, week and sex interface elements.
}
